% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exa.connect.R
\name{exa.connect}
\alias{exa.connect}
\title{Low level method to connect to an EXASOL data base.}
\usage{
exa.connect(host, port = 8563, user = "sys", pwd = "exasol")
}
\arguments{
\item{host}{The host's IP or DNS name to connect to.}

\item{port}{The EXASOL DB port; default: 8563.}

\item{user}{The EXASOL DB user; default: 'sys'.}

\item{pwd}{The EXASOL user password; default: 'exasol'.}
}
\value{
A connection object.
}
\description{
The function connects to an EXASOL database and returns a connection object.

A websockets (RFC 6455) connection is initiated to the EXASOL server, over which a JSON element
is sent to requests the start of the login sequence. Subsequently, a 1024-bit RSA public key is received
from EXASOL DB, to encrypt the user password. The client uses that to build request JSON for login into
EXASOL DB, which also contains some meta information (client name, etc.). EXASOL responds either
with a confirmation containing a session ID and further meta data, or a rejection.
}
\examples{
\dontrun{
    library(exasol6)

    con <- exa.connect("192.168.137.8", 8563, "sys", "exasol")
}
}
\author{
Marcel Boldt <marcel.boldt@exasol.com>
}
