% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exa.readData.R
\name{exa.readData}
\alias{exa.readData}
\title{Low level method to send a query to EXASOL and retrieve a result set.}
\usage{
exa.readData(con, sql, ...)
}
\arguments{
\item{con}{The host's IP or DNS name to connect to.}

\item{sql}{The EXASOL user password; default: 'exasol'.}

\item{...}{Further options to be passed to `as.data.frame()`.}
}
\value{
A connection object.
}
\description{
The function sends an SQL statement via an open connection to an EXASOL DB
and retrieves a result set. The result set is at C++ level fetched via JSON over Websockets,
parsed and handed over to R as a list of vectors, that is then converted to a data.frame.
}
\examples{
\dontrun{
    library(exasol6)

    con <- exa.connect("192.168.137.8", 8563, "R", "sys", "exasol")
    exa.readData(con, "SELECT CURRENT_TIMESTAMP")

    df <- exa.readData(con, "SELECT * FROM R.FLIGHTS LIMIT 100",
                       row.names = c(1:100),
                       stringsAsFactors=FALSE)
}
}
\author{
Marcel Boldt <marcel.boldt@exasol.com>
}
